split projects test into test for shared and test for envs
